requirements.txt
fastavro-1.7.4
google-cloud-pubsub-2.17.1
google-cloud-functions-1.12.0